flexim_py/pyproject.toml
[build-system]
requires = ["maturin>=0.14,<0.15", "grpcio-tools>=1.40.0"]
build-backend = "maturin"


[project]
name = "flexim_py"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]


dependencies = [
    "attrs>=23.1.0",
    "numpy>=1.23",
    "pillow",                 # Used for JPEG encoding
    "pyarrow==14.0.2",
    "typing_extensions>=4.5", # Used for PEP-702 deprecated decorator
    "pydantic>=2.0.0",
    "pandas>=1.3.0",
    "grpcio>=1.40.0",
    "protobuf>=4",
]

[tool]
rye = { dev-dependencies = [
    "pytest>=7.4.4",
    "pip>=24.0",
] }

[tool.maturin]
python-source = "flexim_sdk"
module-name = "flexim_py._flexim_lib"


[tool.ruff]
line-length = 150
indent-width = 4
target-version = 'py312'
